requirements.txt
streamlit
pandas
numpy
requests
plotly
python-dotenv